+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -151,8 +151,8 @@ New checks
   Enforces consistent style for enumerators' initialization, covering three
   styles: none, first only, or all initialized explicitly.
 
-- New :doc:`readability-unnecessary-external-linkage
-  <clang-tidy/checks/readability/unnecessary-external-linkage>` check.
+- New :doc:`misc-use-internal-linkage
+  <clang-tidy/checks/misc/use-internal-linkage>` check.
 
   Detects variable and function can be marked as static.